Crate orml_xtokens

Xtokens Module

Overview

The xtokens module provides cross-chain token transfer functionality, by cross-consensus messages(XCM).

The xtokens module provides functions for

Token transfer from parachains to relay chain.

Token transfer between parachains, including relay chain tokens like DOT, KSM, and parachain tokens like ACA, aUSD.

Interface

Dispatchable functions

transfer: Transfer local assets with given CurrencyId and Amount.

transfer_multiasset: Transfer MultiAsset assets.

Re-exports

pub use module::*;

Modules

module

The module that hosts all the FRAME types needed to add this pallet to a runtime.