Trait orml_xtokens::module::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>;
    type CurrencyId: Parameter + Member + Clone;
    type CurrencyIdConvert: Convert<Self::CurrencyId, Option<MultiLocation>>;
    type AccountIdToMultiLocation: Convert<Self::AccountId, MultiLocation>;
    type SelfLocation: Get<MultiLocation>;
    type MinXcmFee: GetByKey<MultiLocation, Option<u128>>;
    type XcmExecutor: ExecuteXcm<Self::Call>;
    type MultiLocationsFilter: Contains<MultiLocation>;
    type Weigher: WeightBounds<Self::Call>;
    type BaseXcmWeight: Get<Weight>;
    type LocationInverter: InvertLocation;
    type MaxAssetsForTransfer: Get<usize>;
    type ReserveProvider: Reserve;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>

The balance type.

type CurrencyId: Parameter + Member + Clone

Currency Id.

type CurrencyIdConvert: Convert<Self::CurrencyId, Option<MultiLocation>>

Convert T::CurrencyId to MultiLocation.

type AccountIdToMultiLocation: Convert<Self::AccountId, MultiLocation>

Convert T::AccountId to MultiLocation.

type SelfLocation: Get<MultiLocation>

Self chain location.

type MinXcmFee: GetByKey<MultiLocation, Option<u128>>

Minimum xcm execution fee paid on destination chain.

type XcmExecutor: ExecuteXcm<Self::Call>

XCM executor.

type MultiLocationsFilter: Contains<MultiLocation>

MultiLocation filter

type Weigher: WeightBounds<Self::Call>

Means of measuring the weight consumed by an XCM message locally.

type BaseXcmWeight: Get<Weight>

Base XCM weight.

The actually weight for an XCM message is T::BaseXcmWeight + T::Weigher::weight(&msg).

type LocationInverter: InvertLocation

Means of inverting a location.

type MaxAssetsForTransfer: Get<usize>

The maximum number of distinct assets allowed to be transferred in a single helper extrinsic.

type ReserveProvider: Reserve

The way to retreave the reserve of a MultiAsset. This can be configured to accept absolute or relative paths for self tokens